pyproject.toml
[tool.poetry]
name = "sme-agent"
version = "0.1.0"
description = ""
authors = ["Trav <travis.vas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
swekit = "^0.2.8"
langchain = "^0.2.15"
Scaffold = "^0.1.3"
core1 = "^0.0.2"
composio-openai = "^0.5.8"
composio-core = "^0.5.8"
nanoid = "^2.0.0"
prompts = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
